+++ src/callbacksCustom/postPlayerRenderReordered.ts
@@ -1,5 +1,4 @@
 import { isChildPlayer, ModCallbackCustom } from "isaacscript-common";
-import * as fastTeleports from "../features/optional/cutscenes/fastTeleports";
 import * as azazelsRageChargeBar from "../features/optional/quality/azazelsRageChargeBar";
 import * as bloodyLustChargeBar from "../features/optional/quality/bloodyLustChargeBar/bloodyLustChargeBar";
 import * as leadPencilChargeBar from "../features/optional/quality/leadPencilChargeBar";
@@ -20,7 +19,4 @@ function main(player: EntityPlayer) {
   bloodyLustChargeBar.postPlayerRender(player); // 444
   leadPencilChargeBar.postPlayerRender(player); // 444
   azazelsRageChargeBar.postPlayerRender(player); // 699
-
-  // Cutscenes
-  fastTeleports.postPlayerRender(player);
 }